More BUG FIX in Video FIle : this time, I think its much better : 1) I do nothing special in the refresh timer (so there is no jump/ jerk of picture) BUT, 2) when the refresh timer start to be late on presentation time, it tells the Decoding thread to skip the next frame. This way, 1 or 2 frames later, the presentation time stamp to show is 1 frame later, so we can catch up. (+ compilation complain fixed in Rendering Manager XML).
git-svn-id: svn://svn.code.sf.net/p/glmixer/Source/trunk@184 a4deb95c-d308-4ec4-8dc1-6419bbeeb61a

+++ src/RenderingManager.cpp
@@ -835,8 +835,8 @@ QDomElement RenderingManager::getConfiguration(QDomDocument &doc) {
 			specific.appendChild(f);
 
 			QDomElement m = doc.createElement("Marks");
-			m.setAttribute("In", vf->getMarkIn());
-			m.setAttribute("Out", vf->getMarkOut());
+			m.setAttribute("In", (uint) vf->getMarkIn());
+			m.setAttribute("Out", (uint) vf->getMarkOut());
 			specific.appendChild(m);
 
 			QDomElement p = doc.createElement("Play");
@@ -953,7 +953,7 @@ void RenderingManager::addConfiguration(QDomElement xmlconfig) {
 			// if the video file was created successfully
 			if (newSourceVideoFile){
 				// can we open the file ?
-				if ( newSourceVideoFile->open( Filename.text(), marks.attribute("In").toLong(), marks.attribute("Out").toLong() ) ) {
+				if ( newSourceVideoFile->open( Filename.text(), marks.attribute("In").toUInt(), marks.attribute("Out").toUInt() ) ) {
 					// create the source as it is a valid video file (this also set it to be the current source)
 					newsource = RenderingManager::getInstance()->newMediaSource(newSourceVideoFile, depth);
 					if (!newsource)
